move func based view to view.py and add doc string to decorator
