+++ flights/helpers/decorators.py
@@ -3,6 +3,10 @@ from functools import wraps
 
 
 def start_backgroun_job(func):
+    """
+    Start background job on a different thread.
+    """
+
     @wraps(func)
     def decorator(*args, **kwargs):
         t = Thread(target=func, args=args, kwargs=kwargs)
